+++ src/components/DishListComponent.tsx
@@ -79,7 +79,7 @@ export const DishListComponent = ({ dishType }: dishInput) => {
 
   if (isLoading) return <div>Loading...</div>;
   if (error) return <div>Error: {error.message}</div>;
-
+  console.log(data)
   return (
     <>
       <DishesContainer>
